requirements.txt
Mock.GPIO==0.1.9
paho-mqtt==2.1.0
PyYAML==6.0.2
